cleaning up and moving towards unified plotter experience
