+++ src/compas_plotters/__init__.py
@@ -5,8 +5,19 @@ compas_plotters
 
 .. currentmodule:: compas_plotters
 
+
 Classes
 =======
+
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    Plotter
+
+
+Deprecated
+==========
 
 .. autosummary::
     :toctree: generated/
@@ -16,7 +27,6 @@ Classes
     NetworkPlotter
     MeshPlotter
 
-
 """
 
 __version__ = '1.5.0'
@@ -24,14 +34,14 @@ __version__ = '1.5.0'
 from .core import *  # noqa: F401 F403
 from .artists import *  # noqa: F401 F403
 
-from .geometryplotter import GeometryPlotter
-from .plotter import Plotter  # noqa: F401
-from .networkplotter import NetworkPlotter
-from .meshplotter import MeshPlotter
+from ._plotter import BasePlotter  # noqa: F401
+from .networkplotter import NetworkPlotter  # noqa: F401
+from .meshplotter import MeshPlotter  # noqa: F401
+from .geometryplotter import GeometryPlotter  # noqa: F401
+
+from .plotter import Plotter
 
 
 __all__ = [
-    'GeometryPlotter',
-    'NetworkPlotter',
-    'MeshPlotter'
+    'Plotter'
 ]
