+++ src/compas_plotters/artists/polylineartist.py
@@ -2,18 +2,16 @@ from compas_plotters.artists import Artist
 from matplotlib.lines import Line2D
 
 
-__all__ = ['PolylineArtist']
-
-
 class PolylineArtist(Artist):
     """"""
 
     zorder = 1000
 
-    def __init__(self, polyline, draw_points=False, linewidth=1.0, linestyle='solid', color=(0, 0, 0)):
+    def __init__(self, polyline, draw_points=True, linewidth=1.0, linestyle='solid', color=(0, 0, 0)):
         super(PolylineArtist, self).__init__(polyline)
-        self._mpl_polyline = None
-        self._draw_points = draw_points
+        self._mpl_line = None
+        self._point_artists = []
+        self.draw_points = draw_points
         self.polyline = polyline
         self.linewidth = linewidth
         self.linestyle = linestyle
@@ -30,11 +28,14 @@ class PolylineArtist(Artist):
                         linestyle=self.linestyle,
                         color=self.color,
                         zorder=self.zorder)
-        self.mpl_line = self.plotter.axes.add_line(line2d)
+        self._mpl_line = self.plotter.axes.add_line(line2d)
+        if self.draw_points:
+            for point in self.polyline:
+                self._point_artists.append(self.plotter.add(point))
 
     def redraw(self):
         x, y, _ = zip(* self.polyline.points)
-        self.mpl_line.set_xdata(x)
-        self.mpl_line.set_ydata(y)
-        self.mpl_line.set_color(self.color)
-        self.mpl_line.set_linewidth(self.width)
+        self._mpl_line.set_xdata(x)
+        self._mpl_line.set_ydata(y)
+        self._mpl_line.set_color(self.color)
+        self._mpl_line.set_linewidth(self.width)
