+++ src/compas_plotters/networkplotter.py
@@ -1,11 +1,11 @@
 from matplotlib.patches import Circle
-from compas_plotters.plotter import Plotter, valuedict
+from compas_plotters._plotter import BasePlotter, valuedict
 
 
 __all__ = ['NetworkPlotter']
 
 
-class NetworkPlotter(Plotter):
+class NetworkPlotter(BasePlotter):
     """Plotter for the visualisation of COMPAS Networks.
 
     Parameters
